chore: move persistence to writer
