+++ src/common/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod config;
 pub mod dataset;
 pub mod model;
-pub mod persistence;
 pub mod queue;
